refactor(fonts): extract Google Fonts resolution into a shared module
`GoogleFontKey`, `GoogleFontCache` and `downloadFromGoogleFonts` lived inside
`:renderer-android` as `internal`, reachable only from the Robolectric
downloadable-font shadow. The Remote Compose connector needs the same resolution to
honour `RemoteFontFamily.Named("google:…")` in an `.rc` document, and it is a sibling
module — neither depends on the other — so today it cannot see any of it.

A second copy would be the wrong fix. `downloadFromGoogleFonts` encodes a
stage-1/stage-2 contract whose failure mode is subtle and permanent: falling through
to the range query on a *network* error can resolve a key to the family's variable TTF
rather than its static sub-font, and because the result is cached under the same
filename that differently-metricked face then sticks for every later render on the
machine. Two implementations would drift into exactly that.

So it moves to `:data-fonts-google`, a pure-JVM module (okio + okhttp) that both an
Android renderer and a Compose connector can depend on. The whole surface is
`(family, weight, italic) -> File`, which is what keeps it platform-free: `weight` is a
plain CSS numeric rather than a Compose `FontWeight`, because the two callers reach it
from different worlds and neither should adopt the other's type to ask for a file.

No behaviour change. The tests for the moved surface move with it — and earned their
keep immediately: transcribing `slugify` dropped its `.ifEmpty { "font" }` tail, which
would have given every family with no alphanumerics the same `-400.ttf` cache filename.
The test caught it before it could be committed.
